+++ src/settings.ts
@@ -1,10 +1,12 @@
 import { MODULE_ID } from "./module";
 
-type SettingKey = "roguelike" | "plan-ahead" | "unlimited";
+type SettingKey =
+    "roguelike" | "plan-ahead" | "unlimited" | "mark-background-class";
 
 export function getSetting(setting: "plan-ahead"): boolean;
 export function getSetting(setting: "unlimited"): boolean;
 export function getSetting(setting: "roguelike"): boolean;
+export function getSetting(setting: "mark-background-class"): boolean;
 export function getSetting(setting: SettingKey) {
     return game.settings.get(MODULE_ID, setting);
 }
@@ -33,6 +35,16 @@ export function registerSettings() {
     game.settings.register(MODULE_ID, "unlimited", {
         name: "pf2e-skill-issue.settings.unlimited.name",
         hint: "pf2e-skill-issue.settings.unlimited.hint",
+        type: Boolean,
+        scope: "world",
+        default: false,
+        config: true,
+        requiresReload: true,
+    });
+
+    game.settings.register(MODULE_ID, "mark-background-class", {
+        name: "pf2e-skill-issue.settings.mark-background-class.name",
+        hint: "pf2e-skill-issue.settings.mark-background-class.hint",
         type: Boolean,
         scope: "world",
         default: false,