+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [0.8] - 2026-07-18
+
+### Added
+
+- Add icons to show that the skill was improved in the source or by background or class (optional as it doesn't consider rule elements)
+- Add tooltip when hovering over locks to show the reason they are locked
+
 ## [0.7] - 2026-07-17
 
 ### Added